@@ -1,4 +1,4 @@
-package qtx.monitoreo;
+package mx.com.qtx.tdneg.web.monitoreo;
 
 import java.io.IOException;
 import java.util.Date;
@@ -6,14 +6,13 @@ import java.util.Enumeration;
 import java.util.List;
 import java.util.Vector;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.annotation.WebFilter;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
 
 /**
  * Servlet Filter implementation class FiltroMonitoreo
